Pi_Est_with_plot.py
The first Pi_Est.py did not product a plot like the MATLAB file.  I added that.  Instead of merging the two files I uploaded the revisions for an extra example for those new to Python (which includes me as of this upload).

Plus, the plotting feature doesn't work when run from a Windows terminal, but does work when ran in an IDE such as Spyder.  I need to fix that.